Mongodb_fastapi_blog_API-master/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.0
sqlalchemy==2.0.0
asyncpg==0.27.0
alembic==1.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==1.3.1
fastapi-mail==1.2.8
Jinja2==3.1.2
pydantic[email]==1.10.7